feat: update format for the value qty and price, fixed common UI issue

+++ BitMEXOrderBook/Screens/OrderBookScreen/Models/OrderBookEntry.swift
@@ -11,5 +11,5 @@ struct OrderBookEntry: Identifiable, Equatable {
     let id: Int
     let side: String
     let price: Double
-    var size: Int
+    var size: Double
 }
